Allow 'zdns -h' to work even without a lookup module
This is a mild hack. What it does is it adds flag parsing on the failure
path when no valid module is found. The flag parsing will find the "-h"
and print the default usage, terminating the program before hitting
the log.Fatal(). If no -h is supplied it falls through the flag parse
and hits the log.Fatal() as it did previously.