Properly handle all dns RR types returned by miekg
Previously, types unsupported by zdns could be returned via:
-ANY requests
-Resolvers being jerks

If zdns encountered such a record, it would generate an empty answer
as  if..else if... clause lacked an else.

Example:
```
$ echo "example.com" | ./zdns ANY | jq "."
{
  "data": {
    "protocol": "tcp",
    "authorities": [],
    "additionals": [],
    "answers": [
      {
        "min_ttl": 3600,
        "expire": 1209600,
        "ttl": 3580,
        "type": "SOA",
        "name": "example.com.",
        "ns": "sns.dns.icann.org",
        "mbox": "noc.dns.icann.org.",
        "serial": 2016110728,
        "refresh": 7200,
        "retry": 3600
      },
      {},
      {},
      {},
      {
        "answer": "a.iana-servers.net.",
        "name": "example.com",
        "type": "NS",
        "ttl": 86380
      },
      {
        "answer": "b.iana-servers.net.",
        "name": "example.com",
        "type": "NS",
        "ttl": 86380
      },
      {},
      {
        "answer": "2606:2800:220:1:248:1893:25c8:1946",
        "name": "example.com",
        "type": "AAAA",
        "ttl": 86380
      },
      {},
      {
        "answer": "93.184.216.34",
        "name": "example.com",
        "type": "A",
        "ttl": 86380
      },
      {},
      {
        "answer": "v=spf1 -all",
        "name": "example.com",
        "type": "TXT",
        "ttl": 40
      },
      {
        "answer": "$Id: example.com 4415 2015-08-24 20:12:23Z davids $",
        "name": "example.com",
        "type": "TXT",
        "ttl": 40
      },
      {},
      {},
      {},
      {},
      {},
      {}
    ]
  },
  "status": "NOERROR",
  "name": "example.com"
}
```

All empty answers are types miekg understands that zdns does not.

This commit adds a new return struct with 2 fields, 'type' which conforms
to the type field used in the rest of zdns, as well as 'unparsed_rr' which is
the miekg represenation of the data.

The ANY clause was removed, as ANY is not a valid response type (it has no
data) and even if it was returned, it would now get caught by the new else
clause.