Include timestamp in output JSON
Store the the timestamp as RFC3339 string instead of time.Time. This avoids a custom JSONMarshal. Given that the timestamp is generated in the same code block as the call to marshal, it seemed like overkill/overengineering to implement a custom marshaller to keep the true time.Time type for ~3 lines of code. If we need to use the timestamp in the future for anything other than printing, @paul-pearce volunteers to extend it with a custom marshaller.