Fix Zakir's recursive nightmare
Previously, we had issues resolving .TLD1's that had nameservers in
.TLD2.

1) Recursive depth was a complete disaster. Previously the code passed
the max depth in as the starting point, iterating on that. The only
reason anything worked is there was a hard coded value of 10 being used
instead of the actual max.

2) For iterative queries inside of iterative queries, the original
question was being returned instead of the result of the actual
iterative dns queries. This made everything explode