Support for fully contained iterative dns resolutions (no name server necessary) (#94)
* new structure to make recursive not horrid

* basic working iterative lookup for A

* poor but nonzero handling of no records

* handles empty responses correctly

* works for some of the alexa top million just not all

* whos knows whats going on

* starting to provide useful debug messages

* Fix Zakir's recursive nightmare

Previously, we had issues resolving .TLD1's that had nameservers in
.TLD2.

1) Recursive depth was a complete disaster. Previously the code passed
the max depth in as the starting point, iterating on that. The only
reason anything worked is there was a hard coded value of 10 being used
instead of the actual max.

2) For iterative queries inside of iterative queries, the original
question was being returned instead of the result of the actual
iterative dns queries. This made everything explode

* Working on debug statements

* any uses new format

* aaaa to use new format

* moving some debug functions around to have cleaner reading code

* ptr module works

* Remove broken things zakir didn't mean to commit

* Add authority iteration

* ns uses new format

* cname to use new structure

* soa fixed

* spfrr fixed

* txt fixed

* caa to work

* spf and dmarc compile

* woah things compile. I wonder if anything works.

* txt works now

* ns actually looks up ns now

* get cached result

* add cached entry

* Hooks to add caching of glue

* Adds debug entries for the cache

* hand @Pearce a result

* just store the whole damn result

* new interface but no implementation

* Finished framework for feeding results into cache, added layering back

Layering was somehow quashed by zakir's recently pushes. That's back
to handle dealing with glue and poisoning correctly. Also added is
proper handling of results to shove into the cache

* answer inheritence to allow paul's stuff to work

* cleaning up some copy pasted code

* consistent variable naming

* method order

* compiles but is going tg to segfault when it tries to access the global factory

* I broke something. :(

* new fields needed for caching

* cache structure that might actually work

* add when existing item already exists

* broken but close

* Minor code fixes to get caching working further

I've added a temp hack to A records until the framework can be
modified to properly link the Global and Routine factories at the
miekg level

* Fixed a bug in cachehash

* Remove repeated read ulock

* Locks are meant to be released.

* Handle cache expiration properly (nanoseconds -> seconds)

* Fix depth to handle loops

* Adding layer to caching

* Major restructing of caching and answers

I think we're almost there

* Further iterative fixes

* set the response fields to empty lists instead of nil

* Strip additional and authorities from iterative output, if answers exist

* Move to case insensitive comparisons

* Fix authority resolution handling
ns_status -> status

* Fix further case sensitivity

* Iterative timeout

* Fix silly caching bug

* Removed notions of a read lock, fixed a deadlock

Now 100% segfault free!

* slimmer miekg modules

* spf works

* dmarc works

* mx no longer segfaults, but still does not use miekg library correctly

* mxlookup works propertly

* adding pearce as an author

* nslookup works

* removing zone file processing code. I thought we had done this a while ago because it doesn't work

* trying to add tests

* messing around with travis

* messing around with travis

* mxlookup works reasonably now

* no more trailing periods

* no soa periods either

* some tests

* NS test that fail

* Include zdns cmd in unit test failures

* Fix NS queries

Sometimes NS's are answers too!

* Use poison glue, if present

The funny thing about cache poisoning, it only matters if it's in the cache.
It's quite common for domains to return authorities+additionals that are
technically poison. But it's standard to use the poison for that singular
query, and simply not cache it. Example:

ask com, example.com:

authority:
example.com -> foo.net

Additional:
foo.net -> 1.2.3.4

Here, you should not cache foo.com -> 1.2.3.4, but you should use it in the resolution, as they could have easily said:

ask com, example.com:

authority:
example.com -> whatever.com

additonal:
whatever.com -> 1.2.3.4

This is incredibly common across com/net/org

* uncommited changes to tests

* mxlookup tests

* alookup tests

* Remove debug printing

* Handle cache expiration better

If during a cache hit all entries expire, it should really be a cache miss.

* Add miekg in/out wrapper to iterative calls

* Add threadID to the routine interface (ugh)

This is not pretty.

* Logging overhaul

We now log with threadID so we can tease apart verbose log msgs.

This means there's different call to GlobalRoutine logging (where
threadid is unknown) and LookupRoutine logging (where threadid is a
struct variable).

* Better logging, properly register ITERATIVE_TIMEOUT in all cases

* Break out single iteration timeouts, better logging

* Fix iterative timeout msg

* Set udp/tcp timeout properly under iteration, disable rdns on log lines

* Change strings in integration tests to unicode

* Fix nested sorting for MX records

* Further cleanup to integration tests

* Correct dictionary comparisons in integration tests

* Update travis version, fix axfr build issue

I suspect axfr is not in the main build chain at the moment, as
this only errors when doing per-package testing

* Fix integration test bug

* Remove repeated variable declarations

* only one timeout

* comment