Re-expose the bottom-most doLookup behavior (#105)
Long ago we intentionally exposed this so other tools (Ahem, Paul's
stuff) can use this interface. Well, when Zakir and I refactored
miekg for Iterative lookups, we broke this. This once again
exposes the bottom most logic via DoLookupWorker (not a member).

(It also fixes a nameserver bug on an error branch.)