Fix iterative NS lookups with caching
Previously our GetCachedResult function took the original
DNS request type as an argument to disambiguate between authoritative
checks and real requests. This is fundamentally broken in the context
of NS lookups, once the cache was populated, We now explicitly pass
into the function if this is an authoritative check or not.