Add class filter for alookup module to prevent non-A/AAAA records from making it into result set (#167)
* alookup: add test for mixed type responses

* alookup: filter answery by requested type (fix #163)

If a nameserver responds with unexpected RR types those are added to the
IPv4/IPv6 result lists. The alookup module now explicitly filters out
answers of the target type.

Onlythe RR type is checked, the actual valuesa are still passed through.

* alookup: omit possibly remaining CNAME records in output

If a valid A/AAAA record is found and the nameserver responded with an
additional CNAME the domain was added to the result set. Adding an
additional filter solves this problem.

* alookup: refactor inner lookup by spliting candidate and cname sets

Filtering CNAME records twice seems not very clear. To make it explicit
the candidate set not contains only valid type records and the cname set
possible cname candidates. Both sets are passed through the recursion.

* test: A/MX lookup integration tests with v4/v6 only and default (#115)