File I/O Handlers implement interface
We already have an interface for InputHandler and OutputHandler.  This
change removes the registration system and instead relies on the
implementations of the interface methods to do the right thing.  The
GlobalConf object now holds a pointer to an implementation the
interfaces.

    * CLI users could only use the "file" handler type anyway, so this
    option is removed.
    * The registration system (and the associated getters) are removed,
    Since the config holds an implementation, the factory can check that
    the config's InputHandler and OutputHandler are not nil before
    calling their  methods.
    * Some minor changes to Golang's style: for example, avoiding "else"
    after a Fatal keeps the happy-path less indented

This change leaves the functionality of the CLI unmodified (I hope!)
while also making it easier for programs using zdns as a library to
create their own implementations of the I/OHandler interfaces---they can
simply stick them into a znds.GlobalConfig object instead of having to
go through registration.