implement streamhandlers, simplifies I/OHandler interface (#211)
* File I/O Handlers implement interface

We already have an interface for InputHandler and OutputHandler.  This
change removes the registration system and instead relies on the
implementations of the interface methods to do the right thing.  The
GlobalConf object now holds a pointer to an implementation the
interfaces.

    * CLI users could only use the "file" handler type anyway, so this
    option is removed.
    * The registration system (and the associated getters) are removed,
    Since the config holds an implementation, the factory can check that
    the config's InputHandler and OutputHandler are not nil before
    calling their  methods.
    * Some minor changes to Golang's style: for example, avoiding "else"
    after a Fatal keeps the happy-path less indented

This change leaves the functionality of the CLI unmodified (I hope!)
while also making it easier for programs using zdns as a library to
create their own implementations of the I/OHandler interfaces---they can
simply stick them into a znds.GlobalConfig object instead of having to
go through registration.

* Implement streamhandlers, simplifies I/OHandler interface

This change adds streamhandlers which can be use by programs which import
zdns who need other types of input/output handlers beyond the pipes and
disk-files provided by the filehandlers.

* "New" constructors are provided for the File and Stream handler
implementations.  This allows them to keep their filepath (or
reader/writer) private, while also allowing them to be created by
packages which import the handlers.

* The {Input,Output}Handler interfaces are simplified, the Inititalize
method is removed (as are the calls to it in lookup).  This method
wasn't being used by the only existing implementation beyond as a way to
pass the filename to the Handler implementation.  The "New" constructors
fulfill this requirement in a way which is (hopefully) more transparent
and easier to use.

* The "file" package is subsumed by the "iohandlers" package, which
provides the two implementations.

* zdns/main.go is modified to use the File{In,Out}putHandlers

Co-authored-by: Zakir Durumeric <zakird@gmail.com>