fix verbosity=5 runtime bug
Currently at verbosity = 5 we get
"fatal error: concurrent map iteration and map write"
at runtime. This is due to passing a map datastructure
to the logger (which we then iterate over) while another
thread grabs the same structure and edits (deletes due to
timeout) it. Fix is to move the log into the lock. This
sucks for performance but we've already given up on
performance at v=5